morning commit
This commit has had it's current framework from AudioKit's master branch removed, and it should supposedly be using the framework from it's develop branch. I tried using Carthage to get it into my app but it still hasn't resolved the issue.
-- MAIN ISSUE:
The isssue is that the sound (or "startVoice()/stopVoice()") isn't actually starting or stoping the audio. Everything else is in place (or at least should be). I will resume trying to re-import the correct framework after this commit. This commit has no working framework and cannot build or run.